Refactor:   - Move body structure to a seperate file (from base_hpe.py to utils.body.py)   - Move set_threshold() from au_evaluator to parent/base class
